Update on mosaic_lists and dependencies
align - mosaic_lists can now use magnitude cuts on the matching catalogs. The reference catalog is copied as-is, it is not matched with itself anymore. Added the option to run not verbosely. Several warning messages are now avoided.
transforms - A Transform2D object can now read a magnitude offset and use it in the transformation of a catalog.
starlists - New option to add an offset to the magnitude column of a starList.
startables - Explicit option to use sigma-clipping in combining a column of a StarTable. Several warning messages are now avoided.